Implement lodash mixins with typescript and remove unused
